Updates to Logic and more work on the dreamcatcher modules
Updated the Logic for dreamcatcher. Now called terminal.sol, the terminal connects to all external contracts to bring dreamcatcher under one unified logic contract. The way we're building this will allow us to expand the dreamcatcher protocol like a lego block. This will work on matic as its a lot cheaper, if we are to bridge to ethereum we will need to simplify this even further but as polygon is our main, we should take advantage of possible things we can do