Update nTreasury
The treasury is facet of the EIP-2535 and will grant the diamond basic multi sig functionality. The DREAM tokens will be sent here, it will require 'trustee' s to create and sign withdrawal requests. It might seam like there's a bit too many comments but I think this will help with code documentation later on, as everything that needs to be explained is directly in the file.