Reflex python module is a partial success
Compiling times and installing everytime I need to reload localhost takes me 20 minutes because my internet is unusable. Im trying to find our how to fix this, but the module itself works great. I built a navbar in a couple a minutes, subtract 6 years that it took me to load in localhost. In the future this will allow us to ship front end blazingly fast, but for now I have to deal with my personal hell