Back on path
Did lots of rust experimentation, i dont think we have the time to pick it up but itll be a great tool to work with in the future, in the meatime, typescript seams to be where its. I'm in the process of creating the modules to allow for expansive software architecture. It seams like its going to be a lot of work, but I think this will help us work with web3 and build powerful micro services quicker and easier